requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
h11==0.14.0
httpie==3.2.1
idna==3.4
importlib-metadata==4.12.0
Markdown==3.3.7
MarkupPy==1.14
multidict==6.0.2
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.2.0
Pillow==9.2.0
Pygments==2.13.0
PyJWT==2.4.0
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.6.0
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
